Add gray elephant template
In order to guide the user in their task to build
an elephant out of the pieces. We implemented the possibility
to show a template on the right side.
This template can be disabled via the option 'provide_template'
in the config file. For the template we added a gray color to
the color_map.