Integrate Placement Intent into dialog flow
Fill the state 'PlaceSelected' with content.
Furhat can now send two new events: "getHint" and "removeHint".
Those two are used to indicate with a highlight the position
where a piece has to go.
Implement a User Info regarding the previous Action during Placement
and allow user through the use of a new entity 'Back' to reverse
the previous action.
Furhat is now send a new event: "placementSuccessful" when the piece
was locked in the right position on the grid.