Extent the placement functionalities
Add an <Again> block and add an Again intent.
Add a regex to catch a rotation with no extra information.
-> this is only necessary due to a bug with the GrammarEntity
Allow for formulations that include e.g. turn [the cup] to the left.
Add a Demo mode to the game.
Allow for formulations that include e.g. Move it one block to the right.